requirements/common.txt
# Dependencies common to all environments
supervisor==4.2.1
psycopg2-binary==2.8.6
django==3.2.8
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
python-dotenv==0.19.1
django-storages==1.12.2

awscli==1.19.10
